refactor(META): Move to legacy_src
