+++ legacy_src_to_be_removed/components/App/index.scss
