+++ legacy_src_to_be_removed/components/CardItem/index.scss
