+++ legacy_src_to_be_removed/components/Footer/index.js
