+++ legacy_src_to_be_removed/components/Image/component.js
