+++ legacy_src_to_be_removed/components/Image/index.scss
