+++ legacy_src_to_be_removed/components/InfiniteList/index.js
