+++ legacy_src_to_be_removed/components/Loading/index.js
