+++ legacy_src_to_be_removed/literals/index.js
