+++ package.json
@@ -4,58 +4,20 @@
   "private": true,
   "description": "",
   "main": "index.js",
-  "scripts": {
-    "build": "sui-bundler build -C",
-    "co": "npm run lint && git commit",
-    "postbuild": "cp ./src/manifest.json ./public && cp ./src/images/cuantocabron.png ./public",
-    "deploy": "npm run build && firebase deploy",
-    "lint": "npm run lint:js",
-    "lint:js": "sui-lint js --ignore-pattern public --fix",
-    "lint:sass": "sui-lint sass",
-    "start": "sui-bundler dev"
-  },
+  "scripts": {},
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "@schibstedspain/ddd-react-redux": "2",
-    "@schibstedspain/rosetta": "3",
-    "@schibstedspain/sui-image-lazy-load": "1.4.0",
-    "classnames": "2.2.5",
-    "firebase": "4.0.0",
-    "idb-keyval": "2.3.0",
-    "imagesloaded": "4.1.3",
-    "material-ui": "0.18.1",
-    "prop-types": "15.5.10",
-    "react": "15.5.4",
-    "react-dom": "15.5.4",
-    "react-loadable": "3.3.1",
-    "react-router": "4.1.1",
-    "react-router-dom": "4.1.1",
-    "react-tap-event-plugin": "2.0.1",
-    "react-virtual-list": "2.2.0",
-    "recompose": "0.23.4",
-    "reset-css": "2.2.0"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@schibstedspain/sui-bundler": "2",
-    "@schibstedspain/sui-lint": "2"
+    "@s-ui/mono": "1"
   },
   "sui-bundler": {
     "offline": {
       "whitelist": [
         "::all::"
       ]
-    },
-    "vendor": [
-      "@schibstedspain/ddd-react-redux",
-      "@schibstedspain/rosetta",
-      "classnames",
-      "react",
-      "react-dom",
-      "react-redux",
-      "redux"
-    ]
+    }
   },
   "eslintConfig": {
     "extends": [
