--- src/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="manifest" href="/manifest.json">
-    <title>No tengo ganas de trabajar</title>
-
-  </head>
-  <body>
-    <div id="root">
-      Waiting for load JS
-    </div>
-  </body>
-</html>
-
